+++ CameraView.Maui/CameraResolution.cs
@@ -1,0 +1,10 @@
+namespace CameraView.Maui;
+
+public enum CameraResolution
+{
+    Default,
+    Qvga,
+    Vga,
+    Hd720p,
+    Hd1080p
+}